+++ tests/cloudpickle_test.py
@@ -17,6 +17,7 @@ import sys
 import textwrap
 import unittest
 import weakref
+import os
 
 try:
     from StringIO import StringIO
@@ -45,6 +46,9 @@ from cloudpickle.cloudpickle import _find_module, _make_empty_cell, cell_set
 
 from .testutils import subprocess_pickle_echo
 from .testutils import assert_run_python_script
+
+
+_TEST_GLOBAL_VARIABLE = "default_value"
 
 
 class RaiserOnPickle(object):
@@ -435,6 +439,74 @@ class CloudPickleTest(unittest.TestCase):
         # Test dynamic modules when imported back are singletons
         mod1, mod2 = pickle_depickle([mod, mod])
         self.assertEqual(id(mod1), id(mod2))
+
+    def test_load_dynamic_module_in_grandchild_process(self):
+        # Make sure that when loaded, a dynamic module preserves its dynamic
+        # property. Otherwise, this will lead to an ImportError if pickled in
+        # the child process and reloaded in another one.
+
+        # We create a new dynamic module
+        mod = imp.new_module('mod')
+        code = '''
+        x = 1
+        '''
+        exec(textwrap.dedent(code), mod.__dict__)
+
+        # This script will be ran in a separate child process. It will import
+        # the pickled dynamic module, and then re-pickle it under a new name.
+        # Finally, it will create a child process that will load the re-pickled
+        # dynamic module.
+        parent_process_module_file = 'dynamic_module_from_parent_process.pkl'
+        child_process_module_file = 'dynamic_module_from_child_process.pkl'
+        child_process_script = '''
+            import pickle
+            import textwrap
+
+            import cloudpickle
+            from testutils import assert_run_python_script
+
+
+            child_of_child_process_script = {child_of_child_process_script}
+
+            with open('{parent_process_module_file}', 'rb') as f:
+                mod = pickle.load(f)
+
+            with open('{child_process_module_file}', 'wb') as f:
+                cloudpickle.dump(mod, f)
+
+            assert_run_python_script(textwrap.dedent(child_of_child_process_script))
+            '''
+
+        # The script ran by the process created by the child process
+        child_of_child_process_script = """ '''
+                import pickle
+                with open('{child_process_module_file}','rb') as fid:
+                    mod = pickle.load(fid)
+                ''' """
+
+        # Filling the two scripts with the pickled modules filepaths and,
+        # for the first child process, the script to be executed by its
+        # own child process.
+        child_of_child_process_script = child_of_child_process_script.format(
+                child_process_module_file=child_process_module_file)
+
+        child_process_script = child_process_script.format(
+                parent_process_module_file=parent_process_module_file,
+                child_process_module_file=child_process_module_file,
+                child_of_child_process_script=child_of_child_process_script)
+
+        try:
+            with open(parent_process_module_file, 'wb') as fid:
+                cloudpickle.dump(mod, fid)
+
+            assert_run_python_script(textwrap.dedent(child_process_script))
+
+        finally:
+            # Remove temporary created files
+            if os.path.exists(parent_process_module_file):
+                os.unlink(parent_process_module_file)
+            if os.path.exists(child_process_module_file):
+                os.unlink(child_process_module_file)
 
     def test_find_module(self):
         import pickle  # ensure this test is decoupled from global imports
@@ -886,6 +958,40 @@ class CloudPickleTest(unittest.TestCase):
             code = code_template.format(protocol=self.protocol,
                                         clone_func=clone_func)
             assert_run_python_script(textwrap.dedent(code))
+
+    def test_closure_interacting_with_a_global_variable(self):
+        global _TEST_GLOBAL_VARIABLE
+        assert _TEST_GLOBAL_VARIABLE == "default_value"
+        orig_value = _TEST_GLOBAL_VARIABLE
+        try:
+            def f0():
+                global _TEST_GLOBAL_VARIABLE
+                _TEST_GLOBAL_VARIABLE = "changed_by_f0"
+
+            def f1():
+                return _TEST_GLOBAL_VARIABLE
+
+            cloned_f0 = cloudpickle.loads(cloudpickle.dumps(
+                f0, protocol=self.protocol))
+            cloned_f1 = cloudpickle.loads(cloudpickle.dumps(
+                f1, protocol=self.protocol))
+            pickled_f1 = cloudpickle.dumps(f1, protocol=self.protocol)
+
+            # Change the value of the global variable
+            cloned_f0()
+            assert _TEST_GLOBAL_VARIABLE == "changed_by_f0"
+
+            # Ensure that the global variable is the same for another function
+            result_cloned_f1 = cloned_f1()
+            assert result_cloned_f1 == "changed_by_f0", result_cloned_f1
+            assert f1() == result_cloned_f1
+
+            # Ensure that unpickling the global variable does not change its
+            # value
+            result_pickled_f1 = cloudpickle.loads(pickled_f1)()
+            assert result_pickled_f1 == "changed_by_f0", result_pickled_f1
+        finally:
+            _TEST_GLOBAL_VARIABLE = orig_value
 
     @pytest.mark.skipif(sys.version_info >= (3, 0),
                         reason="hardcoded pickle bytes for 2.7")